requirements.txt
transformers
datasets
peft
huggingface_hub
python-dotenv